+++ graphql/mutation.js
@@ -13,25 +13,25 @@ const {
 //MODELS+TYPES
 
 const UserType = require('./types/User');
-const Users = require('../models/user');
+const Users = require('./models/user');
 
 const FarmType = require('./types/Farm');
-const Farms = require('../models/farm');
+const Farms = require('./models/farm');
 
 const LocationType = require('./types/Location');
-const Locations = require('../models/location');
+const Locations = require('./models/location');
 
 const ItemType = require('./types/Item');
-const Items = require('../models/item');
+const Items = require('./models/item');
 
 const UnitType = require('./types/Unit');
-const Units = require('../models/unit');
+const Units = require('./models/unit');
 
 const OrderType = require('./types/Order');
-const Orders = require('../models/order');
+const Orders = require('./models/order');
 
 const InventoryType = require('./types/Inventory');
-const Inventorys = require('../models/inventory');
+const Inventorys = require('./models/inventory');
 
 //MUTATION
